Update patch set 2
Patch Set 2:

I don't know, since multivim is a python project now, all the code style check, unit test and code coverage check is based on python.

But most of the code in this patch is Java. Is there any consideration of the tests mentioned above? Or should we put these java code to another repo?

Patch-set: 2
Label: Code-Review=0