Update patch set 2
Patch Set 2: Code-Review-2

I understand it is for vmware plugin.

My questions is purely for PTL:
- will the Plugin code be subject to design spec too (28973 in this case)?

I am giving -2, assuming the answer is "Yes", Plugin code is also subject to design spec 28973.

If PTL thinks the Plugin code is purely the business of the cloud owner (VMWare in this case), I will agree to PTL's decision. I will remove my -2.

So I need PTL's confirmation that Plugin code is purely the business of cloud owner, and others should not comment on it. Then I will remove my -2.

Patch-set: 2