Update patch set 1
Patch Set 1: Code-Review-2

See my comments of 30133 of design spec, and suggestion of de-scoping it from Beijing Release (R2) and focusing on the the design of a full, normative definition of API and resources in the bigger picture.

See my comments of 35389 regarding waiting for PTL's decision of whether or not Plugin code is purely the business of cloud owner, and other companies should not comment on it.

Thus I give -2 before we resolve 30133 (design), and/or I get clear guidance of 35389.

Patch-set: 1
Label: Code-Review=-2