Update patch set 1
Patch Set 1: Code-Review-1

(5 comments)

Keystone V3 added the concept of domains but it also redefine the concept of tenants for projects. My understanding is that using the concept of "tenant" will be eventually deprecated. Shouldn't be better to stop using tenant_id as variables and use project_id?

Patch-set: 1
Label: Code-Review=-1